+++ CHANGELOG.md
@@ -148,6 +148,7 @@
 - execution contexts can be reused for async calls if they are not active, to avoid constantly requesting memory and creating (heavy) contexts
   - if there is more than 5 contexts, the 6th one will be destroyed once it completes
 - execution contexts are now marked as free to be reused (or deleted) once a value has been computed, without waiting for a call to `await`
+- captures are not renamed anymore by the NameResolutionPass (which used to fully qualify captured names when possible, which isn't desirable: when you capture `&foo`, you expect to be able to use `.foo` not `.module:foo`)
 
 ### Removed
 - removed unused `NodeType::Closure`